feat/like-delete-card

+++ src/utils/api.js
@@ -75,30 +75,30 @@ class Api {
       });
   }
 
-  addLikeCard(id) {
-    return fetch(`${this._baseUrl}cards/likes/${id}`, {
-      method: 'PUT',
-      headers: this._headers,
-    })
-      .then(res => {
-        if (res.ok) {
-          return res.json();
-        }
-        return Promise.reject(`Ошибка: ${res.status}`);
-      });
-  }
-
-  deleteLikeCard(id) {
-    return fetch(`${this._baseUrl}cards/likes/${id}`, {
-      method: 'DELETE',
-      headers: this._headers,
-    })
-      .then(res => {
-        if (res.ok) {
-          return res.json();
-        }
-        return Promise.reject(`Ошибка: ${res.status}`);
-      });
+  changeLikeCardStatus(id, isLiked) {
+    if (isLiked) {
+      return fetch(`${this._baseUrl}cards/likes/${id}`, {
+        method: 'PUT',
+        headers: this._headers,
+      })
+        .then(res => {
+          if (res.ok) {
+            return res.json();
+          }
+          return Promise.reject(`Ошибка: ${res.status}`);
+        });
+    } else {
+      return fetch(`${this._baseUrl}cards/likes/${id}`, {
+        method: 'DELETE',
+        headers: this._headers,
+      })
+        .then(res => {
+          if (res.ok) {
+            return res.json();
+          }
+          return Promise.reject(`Ошибка: ${res.status}`);
+        });
+    }
   }
 
   updateAvatar({link}){
